+++ tests/controlled.spec.ts
@@ -45,21 +45,41 @@ test("hover and blur methods control playback as expected", async ({ page }) => 
   ]);
 });
 
-test("existing event listeners are disabled/re-enabled if the component's controlled state changes", async ({ page }) => {
+test("hover events don't update playback state if the component's controlled state changes", async ({ page }) => {
   await page.goto("/tests/controlled.html");
 
   const hoverVideoPlayer = await page.locator("hover-video-player");
+
+  hoverVideoPlayer.evaluateHandle((el: HoverVideoPlayer) => {
+    (window as any).hoverStartCount = 0;
+    (window as any).hoverEndCount = 0;
+
+    el.addEventListener("hoverstart", () => {
+      (window as any).hoverStartCount++;
+    });
+    el.addEventListener("hoverend", () => {
+      (window as any).hoverEndCount++;
+    });
+  });
 
   // Hover interactions are ignored when controlled
   await expect(hoverVideoPlayer).toHaveJSProperty("controlled", true);
   await hoverVideoPlayer.hover();
   await Promise.all([
+    // The hover/playback state should not be updated, but the hoverstart event did run
     expect(hoverVideoPlayer).not.toHaveAttribute("data-is-hovering", ""),
     expect(hoverVideoPlayer).toHaveAttribute("data-playback-state", "paused"),
+    expect(await page.evaluate(() => (window as any).hoverStartCount)).toBe(1),
+    expect(await page.evaluate(() => (window as any).hoverEndCount)).toBe(0),
   ]);
 
   // Move mouse away so we can try and hover again
   await page.mouse.move(0, 0);
+
+  await Promise.all([
+    expect(await page.evaluate(() => (window as any).hoverStartCount)).toBe(1),
+    expect(await page.evaluate(() => (window as any).hoverEndCount)).toBe(1),
+  ]);
 
   // Disabling the controlled state should allow hover interactions to control playback again
   await hoverVideoPlayer.evaluateHandle((el: HoverVideoPlayer) => { el.controlled = false });
@@ -67,6 +87,8 @@ test("existing event listeners are disabled/re-enabled if the component's contro
   await Promise.all([
     expect(hoverVideoPlayer).toHaveAttribute("data-is-hovering", ""),
     expect(hoverVideoPlayer).toHaveAttribute("data-playback-state", "playing"),
+    expect(await page.evaluate(() => (window as any).hoverStartCount)).toBe(2),
+    expect(await page.evaluate(() => (window as any).hoverEndCount)).toBe(1),
   ]);
 
   // Mouse out is ignored when controlled again
@@ -75,15 +97,25 @@ test("existing event listeners are disabled/re-enabled if the component's contro
   await Promise.all([
     expect(hoverVideoPlayer).toHaveAttribute("data-is-hovering", ""),
     expect(hoverVideoPlayer).toHaveAttribute("data-playback-state", "playing"),
+    expect(await page.evaluate(() => (window as any).hoverStartCount)).toBe(2),
+    expect(await page.evaluate(() => (window as any).hoverEndCount)).toBe(2),
   ]);
 
   // Hover again so we can disable controlled state and hover back out
   await hoverVideoPlayer.hover();
+
+  await Promise.all([
+    expect(await page.evaluate(() => (window as any).hoverStartCount)).toBe(3),
+    expect(await page.evaluate(() => (window as any).hoverEndCount)).toBe(2),
+  ]);
+
   await hoverVideoPlayer.evaluateHandle((el: HoverVideoPlayer) => { el.controlled = false });
   await page.mouse.move(0, 0);
   await Promise.all([
     expect(hoverVideoPlayer).not.toHaveAttribute("data-is-hovering", ""),
     expect(hoverVideoPlayer).toHaveAttribute("data-playback-state", "paused"),
+    expect(await page.evaluate(() => (window as any).hoverStartCount)).toBe(3),
+    expect(await page.evaluate(() => (window as any).hoverEndCount)).toBe(3),
   ]);
 });